EXPExpect: fail test if expectation was not applied until deallocation.
A complete expectation is a an expectation that a matcher was applied
on it. For example `expect(3).to.equal(3)`. But it's very easy to
create an expectation without applying a matcher like `expect(3)` which
is a valid statement. Another, more common instance of an incomplete
expectation is `expect(a).to.beNil` (notice the lack of `()`).

This commit creates a test failure if an EXPExpect object is
deallocated, without any mathcers applied on it, which means the
expectation iscomplete. It's expected that each test is wrapped by an
`@autoreleasepool` so all the expectation would deallocated before the
test ends.